Life-Summarizer-AI/backend/requirements.txt
pandas
sumy
nltk
textblob
matplotlib
cryptography
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
flask-mail==0.9.1
python-dotenv==1.0.0
sumy==0.10.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
pytz
torch
transformers
psycopg2-binary